Proyek Akhir Membangun ETL Pipeline Sederhana/submission-pemda/requirements.txt
pandas~=2.2
requests~=2.32
beautifulsoup4~=4.12
google-auth~=2.36
google-api-python-client~=2.152
pytest~=8.0